Use std::thread instead of calling CreateThread explicitly.
First of all instead of `CreateThread` the function `begintheadex` should be
called, to avoid any kind of leakages in crt, secondly the handle should be
closed. But the most reliable way is to use std::thread and delegate this
routine of handling MS Windows specific stuff to MS team.

Review: https://github.com/adblockplus/adblockplusie/pull/2