Issue #41 - Bring method of determining IE version up to date
Added retrieval functions for the installed IE version, both for the full
version string as well as the numeric major version. Implemented class
Registry_Key. Added unit tests for these.

Remove CAdblockPlusClient::GetIEVersion(), replacing it with the new IE
major version function.

Populate the 'AppInfo::applicationVersion' field. This fixes
https://issues.adblockplus.org/ticket/41

Keep new registry and version functions in their own files. This is in
preparation to use this code as part of the changes needed by
https://issues.adblockplus.org/ticket/404