Issues #1163 - add unit tests for CPluginUserSettings
Add some unit tests for CPluginUserSettings. These are the first unit tests for
the plugin. Added a new project "tests_plugin" for these tests, since a
significant fraction of the plugin code is required to resolve all symbols.

This change is in preparation for refactoring the control flow of
CPluginUserSettings.