Issue #1158 - removed unnecessary definitions and #ifdef/#else/#endif sections
Removed definition of PRODUCT_ADBLOCKPLUS as a command-line definition for the
compiler and assume that it's always defined instead. When this symbol was
defined, so also were SUPPORT_FILTER, SUPPORT_WHITELIST, SUPPORT_DOM_TRAVERSER,
and SUPPORT_FRAME_CACHING. Thus also assume these four symbols are always
defined.

Removed all #ifdef/#else/#endif directives associated with these symbols.
Removed the contents of the associated sections as appropriate.

Assume SUPPORT_CONFIG is never defined. For this symbol, the only actual code
removed was a single static member variable and its initializer. Note that there
is no source file PluginConfig.h, as mentioned in some of the removed sections.