Issue #1186 - Change names that appear in the custom action
The scope of this change set is to rename constant, variable, function, and
class names that appear in the custom action code. These identifiers include
both those present only in the custom action code itself, as well as those
found in the installer library and its unit tests. The name of the entry point
to the custom action library is also present, which appears also in .DEF and
.WXS files.

The scope of this change set does not include identifier that only appear in
the installer, its unit tests, or elsewhere. Nor does it include identifiers
that are not constant, variable, function, or class names. These will be the
subject of future change sets.

Also, there's a removal of some dead code.