Issue 1148 - ABP on chrome blocks ads but IE doesnt
- IHTMLElement::getAttribute is inconvenient to use (see below) but since
Windows XP all versions support IHTMLDOMAttribute and IHTMLElement4.
IHTMLElement::getAttribute sometimes returns not only BSTR or integer but also
null variant or IDispatch which is different from the value written in the html
file. For the cases like <some-tag on-some-event="do();"/> IE returns the value
of 'on-some-event' attribute as 'function(event){\ndo();\n}'. Our filters are
not designed for such transformations, so we need to retrive the value of (body)
of /^.*?\{(body)\}$/. Using IHTMLDOMAttribute and IHTMLElement4 solves the
troubles and simplifies the code.
- fix predecessor processing. We should call `tagName.ToLower();` after the tag
name retrieved. The bug (typo) was we called it before which makes no sense.