Issues #1163, #1173 - refactor CPluginUserSettings
Added exception handlers around the bodies of entry points that need them.
Document the absence of handlers where they're not needed.

Refactored the control flows in the IDispatch::Invoke implementation of
CPluginUserSettings. Eliminate linear searches of string comparisons in both
GetIDsOfNames() and Invoke(). Replaced the linear search (a loop) in
GetIDsOfNames() by an unordered_map lookup. Replaced the linear search (an
if-else sequence) in Invoke() by a switch statement.

The code within the switch statement differs mostly in indentation from the
previous version. Exceptions:
  * Fixed defect within case dispatchID_GetMessage. The type check for the
    VARIANT argument was missing.
  * Fixed a defect in case dispatchID_IsUpdate. The VARIANT_BOOL return value
    had been returning a C++ bool.
  * Swapped terms in yoda expressions.
  * Added braces to conditions.

--HG--
branch : dev-post-1.4