Issue #1762 - Remove smart handle for HMODULE
The return value from LoadLibrary is an HMODULE, which is not a proper handle.
In particular, it shouldn't be closed. Thus the use of a smart handle class to
ensure closing the handle upon destruction is not only unnecessary but
incorrect.

--HG--
extra : rebase_source : 6e5edf8e02b81ca7d6ab0eff746c6d839c1d57c9