Issue #1236 - Update PE metadata in the plugin and the engine
Update the PE (Portable Executable) metadata for the plugin and engine to
current copyright year. Changed to proper company name. Made the descriptions
more precise.

This change set doesn't add the icon as requested by that ticket. The new icon
is under discussion at http://codereview.adblockplus.org/6210609305616384/.
Getting icons compiled in can be either part of that review or follow it.