Issue #404 - Create common library shared between plugin/engine and installer
Create new static library 'common' as a separate project with its own
directory. It uses a separate .gypi include file so that it can be included in
both the plugin/engine solution and in the installer solution. This change set
does not incorporate changes to the installer, because it's not using these
modules yet.

Contents of the library are the registry and IE version classes. Split out the
unit tests for these two modules into separate source files. Refactor the unit
tests to conform to naming conventions. Add one unit test for Registry.