Issue #1486 - Move PassthroughAPP code to its own directory
Move the current PassthroughAPP code into its own directory
'src/plugin/passthroughapp'.

This is a preliminary change. The current code we have for PassthroughAPP is
not fully third-party code at present. It's a combination of an older version
plus our own modifications. After #1484 is done, we can complete #1486 and
move it into 'third_party/passthroughapp'.