Issue #1484 - Remove unused argument from 'WBPassthruSink::OnStart'
The argument 'handled' to 'WBPassthruSink::OnStart' was not referenced.
Removing it brings this member function back to the same argument signature as
defined in its base class 'PassthroughAPP::CInternetProtocolSinkWithSP'.

The only reference to this argument as a local variable 'handled' in
'PassthroughAPP::CustomSinkStartPolicy::OnStart'. Removing this variable
brings the code for this function closer to the original PassthroughAPP code.