Issue #1234 + Clean up - CPluginTab (dead code, nullptr defect fix, etc.)
Replace the last remaining occurrence of 'CString' in "PluginTabBase.cpp".

Remove stub class 'CPluginTab' entirely. Rename base class 'CPluginTabBase'
to 'CPluginTab'. Do not rename files "PluginTabBase.*" to "PluginTab.*". It
would be best to do that in a separate commit because of tool limitationa.

Remove 'CPluginTabBase::m_plugin'. It was initialized in the constructor but
never used. Change the constructor to eliminate unnecessary argument.

Change 'CPluginTabBase::m_filter' from an 'auto_ptr' to an ordinary object. It
had been allocated in the constructor and never changed thereafter. This fixes
a number of defects where the pointer was dereferenced without ever having
been checked for nullptr, neither in the constructor nor at the point of use.

Change all "protected" declarations in 'CPluginTabBase' to "private". There
are no subclasses that reference any affected member.

Changed signature of entry point 'FilterLoader()' to avoid gratuitous need for
'm_filter' to be a public member of 'CPluginTabBase'. There are some remaining
public uses that are not part of this change set.