Issue #3562 - Use 'ToWstring()' to convert BSTR values
Use 'ToWstring()' to convert BSTR values from interfaces. This eliminates
all occurrences of 'SysStringLen()' in the plugin code. (One remains, of
course, in the implementation of 'ToWstring()').

A null BSTR pointer is the same as an empty string. Where needed, change the
logic around BSTR [out] arguments so that these two are treated identically.
This generally means converting first and testing 'empty()' afterwards.