Noissue - Eliminate compile warnings in the plugin
Change the rectangle class to return unsigned values for its height and width
functions. This entails adding guards to check the rectangle invariants.

Add 'NotificationMessage::MoveToCenter()'. This removes code duplication in
three places. In its implementation, explicitly cast coordinates for the
notification message to 'short'. The rectangle coordinates they are derived
from are 'long', but they're screen coordinates, so this operation preserves
the integer value.