No issue - change the lifetime of traverser
It's a part of changes required for issue 119, although it's rather code reorganizing.
The significant change in the behavior is that instead of clearing of the traverser cache we instantiate a new instance of CPluginDomTraverser. We delete the object from "CPluginTab::OnNavigate" and instantiate it when we need it first time.
JIC, it still has the same limitations and does not solve in particular #1201, though it might be a step in direction to #264.