Issue 119 - Fix initialization of CPluginFilter
Exclude the data race on CPluginFilter. Previously it was possible that CPluginFilter is being used by the initializer thread and by the "main" thread (in which OnNavigate and OnDownloadComplete are called).
E.g.
- OnNavigate spawns an initializer thread (t1)
- OnNavigate resets the event and spawns a new initializer thread (t2)
- t1 sets the event
- OnDownloadComplete is called and tries to use CPluginFilter, at the same moment t2 is modifying it.

In addition, now if initialization of CPluginFilter throws an exception the "main" thread won't wait forever (based on EventWithSetter) and on each OnNavigate we clean previously obtained filters and traverser.

add EventWithSetter, it's a workaround similar to the technique std::future/std::promise
add CPluginTab::AsyncPluginFilter which encapsulates auxiliary code for asynchronous CPluginFilter initialization
do LoadHideFilters (parsing) in the constructor of CPluginFilter, thus remove LoadHideFilters
remove CPluginFilter::m_dataPath because it's not used anywhere